requirements.txt
flask==2.3.3
pytesseract==0.3.13
pdf2image==1.17.0
pandas==2.2.3
openpyxl==3.1.5
openai==1.12.0
python-dotenv==0.15.0
Werkzeug==2.3.7 